Change a few things to avoid warnings with gcc. Should fix #2359
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@14529 3db68df2-c116-0410-a063-a993310a9797

+++ source/renderer/WaterManager.cpp
@@ -363,7 +363,7 @@ void WaterManager::CreateSuperfancyInfo(CSimulation2* simulation)
 				// search a 5x5 array with us in the center (do not search me)
 				// much faster since we spiral search and can just stop once we've found the shore.
 				// also everything is precomputed and we get exact results instead.
-				int offset[24] = { -1,1,-m_MapSize,+m_MapSize, -1-m_MapSize,+1-m_MapSize,-1+m_MapSize,1+m_MapSize,
+				size_t offset[24] = { -1,1,-m_MapSize,+m_MapSize, -1-m_MapSize,+1-m_MapSize,-1+m_MapSize,1+m_MapSize,
 					-2,2,-2*m_MapSize,2*m_MapSize,-2-m_MapSize,-2+m_MapSize,2-m_MapSize,2+m_MapSize,
 					-1-2*m_MapSize,+1-2*m_MapSize,-1+2*m_MapSize,1+2*m_MapSize,
 					-2-2*m_MapSize,2+2*m_MapSize,-2+2*m_MapSize,2-2*m_MapSize };
